test: refactor in prep for picture tests
