+++ projects/imgix-ng/src/lib/imgix.component.spec.ts
@@ -1,109 +1,10 @@
-import { Component } from '@angular/core';
-import {
-  render,
-  RenderComponentOptions,
-  screen,
-} from '@testing-library/angular';
 import {
   expectElementToHaveFixedSrcAndSrcSet,
   expectElementToHaveFluidSrcAndSrcSet,
   expectEverySrcsetToMatch,
   expectURLToHaveIxlibParam,
 } from '../test/url-assert';
-import { ImgixConfig } from './imgix-config.service';
-import { NgImgixModule } from './imgix-ng.module';
-import { ImgixComponent } from './imgix.component';
-
-const createComponent = (template: string) => {
-  @Component({
-    template: template,
-  })
-  class TestComponent {}
-
-  return TestComponent;
-};
-
-const createComponentWithTestWrapper = (template: string) => {
-  const testid = genRandId();
-  @Component({
-    template: `<div [attr.data-testid]="testid">${template}</div>`,
-  })
-  class TestComponent {
-    testid: string = testid;
-  }
-
-  return {
-    component: TestComponent,
-    testid,
-    getWrapper: () => screen.getByTestId(testid),
-  };
-};
-
-const renderComponentWithImgixComponent = async (
-  component: Parameters<typeof render>[0],
-  renderOptions?: RenderComponentOptions<ImgixComponent>,
-  config?: ImgixConfig,
-) =>
-  await render(component, {
-    imports: [
-      NgImgixModule.forRoot(
-        config ?? {
-          domain: 'assets.imgix.net',
-        },
-      ),
-    ],
-    excludeComponentDeclaration: true,
-    ...renderOptions,
-  });
-
-let randCounter = 0;
-const genRandId = () => `test-${randCounter++}`;
-
-const renderWithProps = async ({
-  props,
-  config,
-}: {
-  props: Partial<ImgixComponent>;
-  config?: ImgixConfig;
-}) => {
-  const testid = genRandId();
-  const fixture = await render(ImgixComponent, {
-    imports: [
-      NgImgixModule.forRoot(
-        config ?? {
-          domain: 'assets.imgix.net',
-        },
-      ),
-    ],
-    componentProperties: {
-      ...props,
-      htmlAttributes: {
-        'data-testid': testid,
-      },
-    },
-    excludeComponentDeclaration: true,
-  });
-
-  return {
-    fixture,
-    el: screen.getByTestId(testid),
-  };
-};
-
-const renderImgTemplate = async (template: string, config?: ImgixConfig) => {
-  const componentObj = createComponentWithTestWrapper(template);
-  const renderResult = await renderComponentWithImgixComponent(
-    componentObj.component,
-    {
-      declarations: [componentObj.component],
-    },
-    config,
-  );
-
-  return {
-    getComponent: () => componentObj.getWrapper().querySelector('img'),
-  };
-};
+import { renderTemplate } from '../test/util';
 
 describe('Imgix Component', () => {
   it('should combine the domain and path into the src', async () => {
@@ -441,3 +342,10 @@ describe('Imgix Component', () => {
     });
   });
 });
+
+const renderImgTemplate = async (
+  ...args: Parameters<typeof renderTemplate>
+) => {
+  const result = await renderTemplate(...args);
+  return { ...result, getComponent: () => result.getComponent('img') };
+};
