+++ projects/imgix-ng/src/test/url-assert.ts
@@ -1,6 +1,9 @@
 export const stringMatchingFixedSrcSet = (width: number) =>
   jasmine.stringMatching(new RegExp(`.*w=${width}&.*dpr=1.* 1x,`));
 
+export const stringMatchingFixedSrcSetHeight = (heigh: number) =>
+  jasmine.stringMatching(new RegExp(`.*h=${heigh}&.*dpr=1.* 1x,`));
+
 export const expectSrcSetToBeFixed = (
   srcset: string | null | undefined,
   width: number,
